requirements.txt
# Core FastAPI & ASGI Server
fastapi==0.104.1
uvicorn==0.30.1

# Image Processing & YOLO Inference
ultralytics==8.3.25

# File Upload Support
python-multipart==0.0.9

# PIL for image handling
Pillow==10.4.0

# Optional: For logging, metrics, or debugging in prod
# python-dotenv==1.0.1
# loguru==0.7.2

# PyTorch with CUDA 11.8 (for NVIDIA GPU)
torch==2.4.0+cu118
torchvision==0.19.0+cu118
torchaudio==2.4.0+cu118
--extra-index-url https://download.pytorch.org/whl/cu118

# If you need CPU-only (no GPU), use this instead:
# torch==2.4.0+cpu
# torchvision==0.19.0+cpu
# torchaudio==2.4.0+cpu
# --extra-index-url https://download.pytorch.org/whl/cpu